fix: 修复多项bug并完善macOS构建流程
- 重构ServerApp关闭服务器的逻辑，改用条件变量等待客户端断开，避免嵌套事件循环导致的重入问题，移除无用的回调函数
- 修正套接字读取的错误处理逻辑，统一各平台网络接口的返回类型为int，正确处理暂时无数据的场景
- 修复ClientListener中移除未知客户端事件处理器的错误，使用正确的对象调用移除接口
- 修复macOS平台下多线程调用Carbon API导致的崩溃，将键盘布局查询转移到主线程定时执行并缓存结果
- 更新macOS构建打包脚本，修复macdeployqt改写二进制后签名失效的问题，完善构建文档
- 更新.gitignore，新增/build-release忽略目录

+++ src/lib/arch/unix/ArchNetworkBSD.h
@@ -73,7 +73,7 @@ public:
     bool connectSocket(ArchSocket s, ArchNetAddress name) override;
     int pollSocket(PollEntry[], int num, double timeout) override;
     void unblockPollSocket(ArchThread thread) override;
-    size_t readSocket(ArchSocket s, void* buf, size_t len) override;
+    int readSocket(ArchSocket s, void* buf, size_t len) override;
     size_t writeSocket(ArchSocket s, const void* buf, size_t len) override;
     void throwErrorOnSocket(ArchSocket) override;
     bool setNoDelayOnSocket(ArchSocket, bool noDelay) override;
